back to priority

+++ kernel/hilevel.c
@@ -280,9 +280,9 @@ void hilevel_handler_svc(ctx_t* ctx, uint32_t id) {
       uint32_t x = (uint32_t)ctx->gpr[0];
 
       ctx->pc = x;
-      uint32_t a = &main_P3;
-      uint32_t b = &main_P4;
-      uint32_t c = &main_P5;
+      uint32_t a = (uint32_t)&main_P3;
+      uint32_t b = (uint32_t)&main_P4;
+      uint32_t c = (uint32_t)&main_P5;
 
 
       pcb[ newpcb ].ctx.pc = x;
